Add "Reopen all repositories from last session" preference to the general preference section.
Store all the open documents when the app closes and then open them when the app starts again. Defaults to NO, so that there is no change to the existing behavior. Did move things around in applicationDidFinishLaunching: but it should behave the same.

This seems to exacerbate an existing problem where opening a large number of repositories causes the web history view to not load the first commit. Get a webkit error at line 292 of html/views/history/history.js with message "TypeError: Result of expression 'commit' [undefined] is not an object.". Seems to be due to a timing issue with the web view loading the commit before the repository or the PBCommitList has been loaded. This issue exisits in v0.7.1 as well.