KBPopUpToolbarItem: fix icon display when changing size.
IB does not use setImage: (it must set the image to the ivar directly) and so the regularImage and smallImage ivars are not set. This causes the icon to be blank when the user changes the size.

Since the size of the image is set in IB also don't use the default size that KBPopUpToolbarItem uses. Instead allow the toolbar item to resize the image automatically.